+++ src/main/java/pl/pfm/model/category/Category.java
@@ -4,20 +4,20 @@ import pl.pfm.model.parentcategory.ParentCategory;
 
 public class Category {
 
-  private Integer id;
+  private long id;
   private String name;
   private ParentCategory parentCategory;
 
   public Category() {
   }
 
-  public Category(Integer id, String name, ParentCategory parentCategory) {
+  public Category(long id, String name, ParentCategory parentCategory) {
     this.id = id;
     this.name = name;
     this.parentCategory = parentCategory;
   }
 
-  public Integer getId() {
+  public long getId() {
     return id;
   }
